+++ app/admin/bank_transaction.rb
@@ -56,41 +56,4 @@ ActiveAdmin.register BankTransaction do
       params.require(:bank_transaction).permit! #aaahhhhhhhh
     end
   end
-
-  member_action :history do
-    @bank_transaction = BankItem.find(params[:id])
-    @versions = @bank_transaction.versions
-    render "admin/shared/history"
-  end
-
-  action_item :history, only: :show do
-    link_to "Version History", history_admin_bank_item_path(resource)
-  end
-
-  controller do
-    def show
-      @bank_transaction = BankItem.includes(versions: :item).find(params[:id])
-      @versions = @bank_transaction.versions
-      @bank_transaction = @bank_transaction.versions[params[:version].to_i].reify if params[:version]
-      show!
-    end
-  end
-
-  sidebar :versionate, :partial => "admin/shared/version", :only => :show
-
-# See permitted parameters documentation:
-# https://github.com/activeadmin/activeadmin/blob/master/docs/2-resource-customization.md#setting-up-strong-parameters
-#
-# permit_params :list, :of, :attributes, :on, :model
-#
-# or
-#
-# permit_params do
-#   permitted = [:permitted, :attributes]
-#   permitted << :other if resource.something?
-#   permitted
-# end
-
-
-
 end