+++ app/admin/npc_shift.rb
@@ -61,26 +61,4 @@ ActiveAdmin.register NpcShift do
 
     f.actions
   end
-
-  member_action :history do
-    @npc_shift = NpcShift.find(params[:id])
-    @versions = @npc_shift.versions
-    render "admin/shared/history"
-  end
-
-  action_item :history, only: :show do
-    link_to "Version History", history_admin_npc_shift_path(resource)
-  end
-
-  controller do
-    def show
-      @npc_shift = NpcShift.includes(versions: :item).find(params[:id])
-      @versions = @npc_shift.versions
-      @npc_shift = @npc_shift.versions[params[:version].to_i].reify if params[:version]
-      show!
-    end
-  end
-
-  sidebar :versionate, :partial => "admin/shared/version", :only => :show
-
 end